feat(configurations): open the `ConfigurationContent` in a floating card on `HomeSectionDesktop`
