+++ lib/pages/configuration_page/commands/configuration_page_commands.dart
@@ -1,0 +1,3 @@
+import 'package:volume_vault/pages/configuration_page/commands/configuration_page_strategy.dart';
+
+class ConfigurationPageCommands extends ConfigurationPageStrategy {}
